requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.0
fpdf2==2.8.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
gspread==6.1.4
requests==2.32.5
python-dateutil==2.9.0
gunicorn==23.0.0

